Added Search functionality
All recipes from the database are loaded into the RecipeList.js, then search queries that instead of the database directly.
Press enter to search or press the search button.
